Fix(Fall Hacks 2025): Gallery not centered on small screens (#206)
Issue:
On smaller screens, when the gallery images wrapped to a single column,
the images would become left-aligned.

Root cause:
The size of the images were hard coded to be 17.5rem. This meant that if
the gallery wrapped the images would be aligned to the left.

Solution:
Changed the gallery to use grid instead. The images will now attempt to
grow to fill the column, unless the column has 17.5rem or higher of
horizontal space. At that point, a new column will be made and the
row will accommodate another image.